% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2c.R
\name{c2c}
\alias{c2c}
\title{c2c}
\usage{
c2c(
  end_date = NA_character_,
  index = "IPC",
  level = "G",
  data = NULL,
  x = NA_character_,
  start_date = NA_character_
)
}
\arguments{
\item{end_date}{fecha final para la conversion "yyyy-mm-dd"}

\item{index}{indice a utilizar para convertir (IPC, IPAB)}

\item{level}{por defecto, general}

\item{data}{data frame con los montos a convertir}

\item{x}{variable de data con los valores a convertir}

\item{start_date}{variable de data con la fecha correspondiente a los valores "yyyy-mm-dd"}
}
\value{
data.frame
}
\description{
convierte de pesos uruguayos corrientes a constantes
}
\examples{
#'
}
