% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotizaciones}
\alias{cotizaciones}
\title{A dataset containing the currency exchange between pesos uruguayos and various foreign currency}
\format{
A data frame with more than 8000 rows and 11 variables:
\describe{
\item{fecha}{date from 1999 to today}
\item{USD compra}{}
\item{USD venta}{}
\item{USD eBROU compra}{}
\item{USD eBROU venta}{}
\item{EUR compra}{}
\item{EUR venta}{}
\item{ARS compra}{}
\item{ARS venta}{}
\item{BRL compra}{}
\item{BRL venta}{}
}
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
cotizaciones
}
\description{
A dataset containing the currency exchange between pesos uruguayos and various foreign currency
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{divisas}},
\code{\link{ipc_base2010}},
\code{\link{ipc_base2022}},
\code{\link{presupuesto_2021}},
\code{\link{urls_ine}},
\code{\link{urls_opp}}
}
\concept{dataset}
\keyword{datasets}
