% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_converter.R
\name{currency_converter}
\alias{currency_converter}
\title{currency_converter}
\usage{
currency_converter(
  data = NULL,
  x = NA_character_,
  money_from = "UYU",
  money_to = "USD",
  value = "buy",
  ebrou = FALSE,
  base_date = Sys.Date() - 1
)
}
\arguments{
\item{data}{data.frame}

\item{x}{variable numerica a convertir}

\item{money_from}{por defecto, pesos uruguayos}

\item{money_to}{por defecto, dolares estadounidenses}

\item{value}{'buy', 'sell', 'mean'}

\item{ebrou}{si es TRUE utiliza la cotizacion eBROU}

\item{base_date}{fecha de la cotizacion a utilizar}
}
\value{
data.frame
}
\description{
currency_converter
}
\examples{
#' 
}
