% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divisas}
\alias{divisas}
\title{A dataset containing the currency code and name}
\format{
A data frame with more than 8000 rows and 11 variables:
\describe{
\item{codigo}{ISO 4217 que define con 3 letras el codigo de cada divisa}
\item{moneda}{nombre de la divisa}
}
}
\source{
\url{https://es.wikipedia.org/}
}
\usage{
divisas
}
\description{
A dataset containing the currency code and name
}
\seealso{
Other dataset: 
\code{\link{cotizaciones}},
\code{\link{ipc_base2010}},
\code{\link{ipc_base2022}},
\code{\link{presupuesto_2021}},
\code{\link{urls_ine}},
\code{\link{urls_opp}}
}
\concept{dataset}
\keyword{datasets}
