% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency.R
\name{format_currency_data}
\alias{format_currency_data}
\title{format_currency_data}
\usage{
format_currency_data(write = FALSE, path = tempdir())
}
\arguments{
\item{write}{por defecto no guarda el objeto en un archivo}

\item{path}{por defecto, una carpeta temporal}
}
\value{
data.frame
}
\description{
format_currency_data
}
\examples{
#'
}
\author{
Guzman Lopez
}
