% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budget.R
\name{get_budget}
\alias{get_budget}
\title{get_budget}
\usage{
get_budget(year = 2021, level = "national", folder = tempdir(), toR = TRUE)
}
\arguments{
\item{year}{anio del credito presupuestal}

\item{level}{a nivel nacional ('national') o subnacional ('subnational')}

\item{folder}{carpeta donde descargar el archivo}

\item{toR}{convertir a formato RData}
}
\value{
file
}
\description{
get_budget
}
\examples{
\donttest{
get_budget(year = 2021, toR = FALSE)
}
}
