% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency.R
\name{get_currency_exchange}
\alias{get_currency_exchange}
\title{get_currency_exchange}
\usage{
get_currency_exchange(path = tempdir(), write = FALSE)
}
\arguments{
\item{path}{path to download files and processed data}

\item{write}{FALSE to not save data}
}
\value{
data.frame
}
\description{
get_currency_exchange
}
\examples{
\donttest{
get_currency_exchange()
}
}
