% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency.R
\name{get_currency_exchange_data}
\alias{get_currency_exchange_data}
\title{get_currency_exchange_data}
\usage{
get_currency_exchange_data()
}
\value{
data.frame
}
\description{
get_currency_exchange_data
}
\examples{
\donttest{
get_currency_exchange_data()
}
}
\author{
Guzman Lopez
}
