% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipc}
\alias{get_ipc}
\title{get_ipc}
\usage{
get_ipc(path = tempdir())
}
\arguments{
\item{path}{por defecto, una carpeta temporal donde se descarga una planilla}
}
\value{
data.frame
}
\description{
get_ipc
}
\examples{
\donttest{
get_ipc()
}
}
