% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_base2022}
\alias{ipc_base2022}
\title{A dataset containing the IPC base 2022}
\format{
A data frame with 1024 rows and 8 variables:
\describe{
\item{fecha}{date from 1937 to now}
\item{indice}{IPC}
\item{mensual}{mensual value of IPC}
\item{trimestre}{three-month period value of IPC}
\item{cuatrimestre}{four-month period value of IPC}
\item{semestre}{six-month period value of IPC}
\item{acum_ano}{acumulated IPC}
\item{acum_12_meses}{acumulated IPC last 12 month}
}
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ipc_base2022
}
\description{
A dataset containing the IPC base 2022
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cotizaciones}},
\code{\link{divisas}},
\code{\link{ipc_base2010}},
\code{\link{presupuesto_2021}},
\code{\link{urls_ine}},
\code{\link{urls_opp}}
}
\concept{dataset}
\keyword{datasets}
