% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presupuesto_2021}
\alias{presupuesto_2021}
\title{A dataset containing the National Bugdet}
\format{
A data frame with more than 38000 rows and 29 variables:
\describe{
\item{anio}{}
\item{organismo_nombre}{}
\item{ue_nombre}{}
\item{grupo_ues_nombre}{}
\item{ap_nombre}{}
\item{programa_nombre}{}
\item{proyecto_codigo}{}
\item{proyecto_nombre}{}
\item{financ_nombre}{}
\item{fuente_financ_nombre}{}
\item{tipo_gasto_nombre}{}
\item{grupo_nombre}{}
\item{subgrupo_nombre}{}
\item{odgyaux_codigo}{}
\item{odgyaux_nombre}{}
\item{ods_nombre}{}
\item{apertura}{}
\item{credito}{}
\item{ejecutado}{}
\item{organismo_codigo}{}
\item{ue_codigo}{}
\item{grupo_ues_codigo}{}
\item{ap_codigo}{}
\item{programa_codigo}{}
\item{financ_codigo}{}
\item{fuente_financ_codigo}{}
\item{grupo_codigo}{}
\item{subgrupo_codigo}{}
\item{ods_codigo}{}
}
}
\source{
\url{http://www.transparenciapresupuestaria.opp.gub.uy/}
}
\usage{
presupuesto_2021
}
\description{
A dataset containing the National Bugdet
}
\seealso{
Other dataset: 
\code{\link{cotizaciones}},
\code{\link{divisas}},
\code{\link{ipc_base2010}},
\code{\link{ipc_base2022}},
\code{\link{urls_ine}},
\code{\link{urls_opp}}
}
\concept{dataset}
\keyword{datasets}
