% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_off}
\alias{round_off}
\title{round_off}
\usage{
round_off(x, digits = 0)
}
\arguments{
\item{x}{vector}

\item{digits}{number of digits}
}
\value{
vector
}
\description{
round_off
}
\examples{
round_off(5.8888)
}
