% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budget.R
\name{summary_budget}
\alias{summary_budget}
\title{summary_budget}
\usage{
summary_budget(data = budgetuy::presupuesto_2021, level = "org")
}
\arguments{
\item{data}{data.frame}

\item{level}{area, organismo}
}
\value{
data.frame
}
\description{
summary_budget
}
\examples{
#' 
}
