% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls_ine}
\alias{urls_ine}
\title{A dataset containing the urls of INE datasets}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{indice}{date from 2011 to 2022}
\item{url}{level of the budget}
\item{nota}{url for data download}
}
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
urls_ine
}
\description{
A dataset containing the urls of INE datasets
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cotizaciones}},
\code{\link{divisas}},
\code{\link{ipc_base2010}},
\code{\link{ipc_base2022}},
\code{\link{presupuesto_2021}},
\code{\link{urls_opp}}
}
\concept{dataset}
\keyword{datasets}
