% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls_opp}
\alias{urls_opp}
\title{A dataset containing the urls of OPP datasets and metadata}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{yy}{date from 2011 to 2022}
\item{cpd_csv}{url for data download}
\item{dic}{url for metadata download}
}
}
\source{
\url{http://www.transparenciapresupuestaria.opp.gub.uy/}
}
\usage{
urls_opp
}
\description{
A dataset containing the urls of OPP datasets and metadata
}
\details{
Disclaimer: This script is not an official OPP product.
Aviso: El script no es un producto oficial de OPP.
}
\seealso{
Other dataset: 
\code{\link{cotizaciones}},
\code{\link{divisas}},
\code{\link{ipc_base2010}},
\code{\link{presupuesto_2021}}
}
\concept{dataset}
\keyword{datasets}
