% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{values_format}
\alias{values_format}
\title{values_format}
\usage{
values_format(data = NULL, x = NA_character_, to = "thousand")
}
\arguments{
\item{data}{data.frame}

\item{x}{variable a convertir}

\item{to}{a miles ('thousand') o millones ('million')}
}
\value{
data.frame
}
\description{
values_format
}
\examples{
#' 
}
