% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2c.R
\name{c2c}
\alias{c2c}
\title{c2c}
\usage{
c2c(
  to = "constant",
  base_month = NULL,
  base_year = NULL,
  index = "IPC",
  level = "G",
  data = NULL,
  x = NA_character_,
  y = NA_character_,
  m = NA_character_
)
}
\arguments{
\item{to}{por defecto, 'constant' para convertir a pesos constantes}

\item{base_month}{mes base}

\item{base_year}{anio base}

\item{index}{indice a utilizar para convertir (IPC, IPAB)}

\item{level}{por defecto, general}

\item{data}{data frame con los montos a convertir}

\item{x}{variable de data con los montos}

\item{y}{variable de data con el anio}

\item{m}{variable de data con el mes}
}
\value{
data.frame
}
\description{
convierte de pesos uruguayos corrientes a constantes
}
\examples{
#'
}
