% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkdate}
\alias{checkdate}
\title{checkdate}
\usage{
checkdate(date = NA_character_)
}
\arguments{
\item{date}{character date 'yyyy-mm-dd'}
}
\value{
TRUE if date is valid date
}
\description{
checkdate
}
\examples{
#' 
}
